Add booking history page, enhance forgot password flow, improve login and registration forms, and update user profile layout
- Created a new booking history page with filters, search functionality, and pagination.
- Enhanced the forgot password page to include multiple steps for email verification and password reset.
- Improved the login page to display success and error messages.
- Updated the registration form to include a phone number field.
- Revamped the user profile page with tabs for personal information, booking history, and account settings.
- Added responsive design elements and improved styling across various templates.

+++ src/main/java/com/viecinema/moviebooking/controller/MovieControler.java
@@ -33,7 +33,7 @@ public class MovieControler {
     private ScreenService screenService;
 
     @GetMapping("/movie-detail")
-    public String getMovieDetailPage(@RequestParam("id") Integer id, Model model) {
+    public String getMovieDetailPage(@RequestParam("id") Long id, Model model) {
         MovieDTO movie = movieService.getMovieById(id);
         List<Showtime> showtimes = showtimeService.getShowtimesByMovieId(movie.getId());
         List<List<String>> formattedShowtimeDates = getFormattedShowtimeDates(showtimes);
