+++ src/main/java/com/viecinema/moviebooking/repository/MovieRepository.java
@@ -9,8 +9,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface MovieRepository extends JpaRepository<Movie, Integer> {
-    Optional<Movie> findByTmdbID(int tmdbID);
+public interface MovieRepository extends JpaRepository<Movie, Long> {
+    Optional<Movie> findByTmdbID(Long tmdbID);
     List<Movie> findByReleaseDateBetween(LocalDate startDate, LocalDate endDate);
 
     List<Movie> findByReleaseDateAfter(LocalDate date);
