+++ src/main/java/com/viecinema/moviebooking/repository/TrailerRepository.java
@@ -5,9 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
-public interface TrailerRepository extends JpaRepository<Trailer, Integer> {
-    List<Trailer> findByMovieId(Integer id);
+public interface TrailerRepository extends JpaRepository<Trailer, Long> {
+    List<Trailer> findByMovieId(Long id);
 }
